refactor(bezier-icons): separate getIconNamesByStatus function and test code of getChangeLog
